{{cookiecutter.repo_name}}/requirements/local.txt
-r base.txt
django-debug-toolbar
django-extensions